+++ src/main/java/careerfestival/career/repository/CommentRepository.java
@@ -8,7 +8,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {
@@ -17,10 +16,7 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
     List<Comment> findByOrderNumber(Long orderNumber);
 
     @Query("SELECT COALESCE(MAX(c.orderNumber), 0) FROM Comment c WHERE c.parent IS NULL")
-    Integer findMaxOrderNumber();
-
-
-
-
+    Long findMaxOrderNumber();
+    
 
 }
